requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pdfminer.six
scikit-learn
joblib
nltk
openai
python-dotenv
pandas